+++ python/lib/sift_client/util/util.py
@@ -7,6 +7,7 @@ if TYPE_CHECKING:
         AnnotationsAPIAsync,
         AssetsAPIAsync,
         CalculatedChannelsAPIAsync,
+        CampaignsAPIAsync,
         ChannelsAPIAsync,
         DataExportAPIAsync,
         DataImportAPIAsync,
@@ -38,6 +39,9 @@ class AsyncAPIs(NamedTuple):
 
     calculated_channels: CalculatedChannelsAPIAsync
     """Instance of the Calculated Channels API for making asynchronous requests."""
+
+    campaigns: CampaignsAPIAsync
+    """Instance of the Campaigns API for making asynchronous requests."""
 
     channels: ChannelsAPIAsync
     """Instance of the Channels API for making asynchronous requests."""